Fix issues in CSS-Editor
If Hyphenator.js was run on elements that have been loaded by script
sometime an error occurred saying something like
`IndexSizeError: DOM Exception 1: Index or size was negative, or greater than the allowed value.`

This Error was caused by the repeated run where a change of CSS was registered for each element.
When hyphenation was done, these changes where cleared in CSS but the list of changes still pointed to
these changes (that didn't exist anymore).

This issue is fixed now (and the code is shorter).

Other fixes:
  - add classes to an existing styleSheet (don't create a new one)
  - do not clear visibility that has been set by the user (this has to be done in `onhyphenationdonecallback`) see test92.html
  - test92.html is updated to work with these changes