Add aprtial support for compound characters
Characters can be compound (e.g. ñ = n + tilde).
To handle this we decompose characters in lo.specialChars and add
those to the regex finding words.
Decomposition is done with String().normalize("NFD") wich isn't
yet supported on all browsers (works on Chrome and FF, but not in Safari and IE).